+++ packages/crypto/src/browser/index.ts
@@ -1,7 +1,9 @@
 import type { CryptoApi } from '../types';
 
-import { bufferFromString, stringFromBuffer, decrypt, encrypt, keyFromPassword } from './aes-ctr';
+import { decrypt, encrypt, keyFromPassword } from './aes-ctr';
+import { bufferFromString } from './bufferFromString';
 import { randomBytes } from './randomBytes';
+import { stringFromBuffer } from './stringFromBuffer';
 
 const api: CryptoApi = {
   bufferFromString,
